doc/afl-cheatsheet: add libclang_rt.builtins-* instructions.
If you're building stuff on x86 with clang + UBSan, you're likely to run
into a linker error similar to this:

    foobar.c:(.text+0xf00b): undefined reference to `__mulodi4'

This bug is tracked here:
https://bugs.llvm.org/show_bug.cgi?id=17693

The work-around is to add the libclang_rt.builtins-i686.a library
explicitly to the linker command-line.

The commands added in this commit ask afl-clang-fast for the exact path
of this file. The former requires clang ≥ 4.0; the latter is ugly, but
works with any clang.